ChatElement 분리
