+++ lib/Words.dart
@@ -55,13 +55,13 @@ class Words extends ChangeNotifier{
     '스시',
     '카레',
     '팟타이',
-    '초코 퐁듀',
+    '초코퐁듀',
     '햄치즈오믈렛',
     '오코노미야끼',
     '양꼬치',
     '스콘',
     '핫초콜릿',
-    '비프 웰링턴',
+    '비프웰링턴',
     '마카롱',
     '와플',
     '도넛',
@@ -79,7 +79,7 @@ class Words extends ChangeNotifier{
     '츄러스',
     '파운드케이크',
     '수플레',
-    '아이스크림 콘',
+    '아이스크림콘',
     '호두파이',
     '모나카',
     '커피',
